Codechange: Move NewGRF string mapping to own file.
