+++ src/newgrf/CMakeLists.txt
@@ -2,4 +2,6 @@ add_files(
     newgrf_bytereader.cpp
     newgrf_bytereader.h
     newgrf_internal.h
+    newgrf_stringmapping.cpp
+    newgrf_stringmapping.h
 )
